refactor: rename static/ directory to derive/
