+++ packages/core/src/derive/__tests__/condenseVars.test.ts
