+++ packages/core/src/derive/__tests__/declareVar.test.ts
