+++ packages/core/src/derive/__tests__/extractVars.test.ts
