+++ packages/core/src/derive/__tests__/indexVars.test.ts
