+++ packages/core/src/derive/condenseVars.ts
