+++ packages/core/src/derive/declareVar.ts
