+++ packages/core/src/derive/extractVars.ts
