+++ packages/core/src/derive/indexVars.ts
