+++ packages/core/src/derive/utils/__tests__/sanitizeVar.test.ts
