+++ packages/core/src/derive/utils/regex.ts
