+++ packages/core/src/derive/utils/sanitizeVar.ts
