+++ packages/core/src/derive/utils/traverseHelpers.ts
